refactor(schemas): migrate Phase 24 endpoint schemas — 2 files, 7 classes (#6042)
- integration_cloud.py (3) → schemas_workflows.py
  rename ConnectionTestRequest → CloudConnectionTestRequest
- integration_monitoring.py (4) → schemas_workflows.py
  rename ConnectionTestRequest → MonitoringConnectionTestRequest
  Pydantic v1 @validator → v2 @field_validator + @classmethod
  Cross-field validation: values: Dict[str, Any] → info: ValidationInfo via info.data

Cleanup: removed dead-code imports (ResourceListRequest, MonitorCreateRequest)
